// frontend/client/src/types.tsx
export type BroadcastState = {
  isOnline: boolean;
  broadcast?: {
    likeCount: number;
    id: number;
    title: string;
    startAt: string;
    listenerPeakCount: number;
  };
};